CS-5833: Move logic out of the init file, add link pointing to a CodeScene instance
